Write merge function to update compile diagnostics
Originally, we wholesale overwrote our diagnostics each time they were
generated - but that means we lose info in our symbol table when there
are errors. So now, instead of overwriting, we merge the incoming
results with the old ones. We always update the errors, but we only
update the symbol table if there aren't errors.

We also needed to separate cppfront errors from cpp1 errors so cpp1
errors don't break the algorithm